% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrynames.R
\name{howmany}
\alias{howmany}
\title{Count Number of Countries Mentioned in Text}
\usage{
howmany(text, ignore_case = TRUE)
}
\arguments{
\item{text}{A character string containing the text to search for country names}

\item{ignore_case}{Logical. If TRUE, performs case-insensitive matching. Default is TRUE.}
}
\value{
An integer representing the number of unique countries found
}
\description{
Count Number of Countries Mentioned in Text
}
\examples{
howmany("I visited France and Germany last summer")
howmany("The conference will be held in Japan")
}
